Ros2 lynx merge

+++ husarion_ugv_manager/src/safety_manager_node.cpp
@@ -24,7 +24,6 @@
 #include <utility>
 #include <vector>
 
-#include "ament_index_cpp/get_package_share_directory.hpp"
 #include "behaviortree_ros2/ros_node_params.hpp"
 #include "rclcpp/rclcpp.hpp"
 
@@ -138,7 +137,7 @@ std::map<std::string, std::any> SafetyManagerNode::CreateSafetyInitialBlackboard
   const double cpu_fan_off_temp = this->params_.cpu.temp.fan_off;
   const double driver_fan_on_temp = this->params_.driver.temp.fan_on;
   const double driver_fan_off_temp = this->params_.driver.temp.fan_off;
-  const double fan_turn_off_timeout = this->params_.fan_turn_off_timeout;
+  const float fan_turn_off_timeout = static_cast<float>(this->params_.fan_turn_off_timeout);
 
   const std::map<std::string, std::any> safety_initial_bb = {
     {"CPU_FAN_OFF_TEMP", cpu_fan_off_temp},
